posts/requirements.txt
Flask
flask-restful
flask-mysql
PyJWT
Pillow
requests
flask-cors
boto3